Moving more operators into individual files.
