+++ __init__.py
@@ -32,6 +32,25 @@ from .ops.imports import(
 	LoadSetFile,
 	LoadAnimFile
 )
+from .ops.materials import(
+	AutoAssignTextures,
+	ToPrincipledBsdf,
+	UpdateMaterials,
+	MatToAssetLibrary
+)
+from .ops.mesh import(
+	StrippifyTest
+)
+from .ops.object import(
+	ArmatureFromObjects
+)
+from .ops.textures import(
+	AddTextureSlot,
+	RemoveTextureSlot,
+	MoveTextureSlot,
+	ClearTextureList,
+	AutoNameTextures
+)
 
 # meta info
 bl_info = {
@@ -73,664 +92,6 @@ class TOPBAR_MT_SA_export(bpy.types.Menu):
 		layout.operator("export_scene.sa1lvl")
 		layout.operator("export_scene.sa2lvl")
 		layout.operator("export_scene.sa2blvl")
-
-# operators
-class StrippifyTest(bpy.types.Operator):
-	'''An operator for test-strippifying a model'''
-	bl_idname = "object.strippifytest"
-	bl_label = "Strippify (testing)"
-	bl_description = "Strippifies the active model object"
-
-	doConcat: BoolProperty(
-		name = "Concat",
-		description="Combines all strips into one big strip",
-		default=False
-		)
-
-	doSwaps: BoolProperty(
-		name = "Utilize Swapping",
-		description = "Utilizes swapping when creating strips, which can result in a smaller total strip count",
-		default = False
-		)
-
-	raiseTopoError: BoolProperty(
-		name = "Raise Topo Error",
-		description = "Raise Topo Error if any occur",
-		default = False
-		)
-
-	def invoke(self, context, event):
-		wm = context.window_manager
-		return wm.invoke_props_dialog(self)
-
-	def execute(self, context):
-		import os
-		os.system("cls")
-		obj = context.active_object
-		if obj is None or not isinstance(obj.data, bpy.types.Mesh):
-			print("active object not a mesh")
-			return {'FINISHED'}
-
-		ob_for_convert = obj.original
-		me = ob_for_convert.to_mesh(preserve_all_data_layers=True)
-		common.trianglulateMesh(me)
-
-		# creating the index list
-		indexList = [0] * len(me.polygons) * 3
-
-		for i, p in enumerate(me.polygons):
-			for j, li in enumerate(p.loop_indices):
-				indexList[i * 3 + j] = me.loops[li].vertex_index
-
-		# strippifying it
-		from . import strippifier
-		try:
-			indexStrips = strippifier.Strippify(indexList, doSwaps = self.doSwaps, concat = self.doConcat, raiseTopoError=self.raiseTopoError)
-		except strippifier.TopologyError as e:
-			self.report({'WARNING'}, "Topology error!\n" + str(e))
-			return {'CANCELLED'}
-
-
-		empty = bpy.data.objects.new(obj.data.name + "_str", None)
-		context.collection.objects.link(empty)
-		for i, s in enumerate(indexStrips):
-			# making them lists so blender can use them
-
-			verts = dict()
-			for p in s:
-				verts[p] = me.vertices[p].co
-
-			indexList = list()
-			firstTri = [s[i] for i in range(3)]
-
-			rev = len(set(firstTri)) == 3
-			for j in range(0 if rev else 1, len(s)-2):
-				if rev:
-					p = [s[j+1], s[j], s[j+2]]
-				else:
-					p = [s[j], s[j+1], s[j+2]]
-				if len(set(p)) < 3:
-					print("something wrong...")
-				indexList.append(p)
-				rev = not rev
-
-			keys = list(verts.keys())
-			indexList = [[keys.index(i) for i in l] for l in indexList]
-			verts = list(verts.values())
-
-			mesh = bpy.data.meshes.new(name = obj.data.name + "_str_" + str(i))
-			mesh.from_pydata(verts, [], indexList)
-			meObj = bpy.data.objects.new(mesh.name, object_data = mesh)
-			context.collection.objects.link(meObj)
-			meObj.parent = empty
-
-		return {'FINISHED'}
-
-class ArmatureFromObjects(bpy.types.Operator):
-	'''Generates an armature based on the selected node and its child hierarchy'''
-	bl_idname = "object.armaturefromobjects"
-	bl_label = "Armature from objects"
-	bl_description = "Generate an armature from object. Select the parent of all objects, which will represent the root"
-
-	rotMode: EnumProperty(
-		name="Rotation Mode",
-		description="The rotation mode of each bone in the generated armature",
-		items=( ('QUATERNION', "Quaternion (WXYZ)", "No Gimbal Lock."),
-				('XYZ', "XYZ Euler", "XYZ Rotation Order - prone to Gimbal Lock (default)."),
-				('XZY', "XZY Euler", "XZY Rotation Order - prone to Gimbal Lock."),
-				('YXZ', "YXZ Euler", "YXZ Rotation Order - prone to Gimbal Lock."),
-				('YZX', "YZX Euler", "YZX Rotation Order - prone to Gimbal Lock."),
-				('ZXY', "ZXY Euler", "ZXY Rotation Order - prone to Gimbal Lock."),
-				('ZYX', "ZYX Euler", "ZYX Rotation Order - prone to Gimbal Lock.") ),
-		default='QUATERNION'
-		)
-
-	mergeModel: BoolProperty(
-		name="Merge Meshes",
-		description="Generates a single mesh object instead of keeping the single objects",
-		default=False
-		)
-
-	def invoke(self, context, event):
-		wm = context.window_manager
-		return wm.invoke_props_dialog(self)
-
-	@classmethod
-	def poll(cls, context):
-		return len(context.selected_objects) > 0
-
-	@classmethod
-	def addChildren(cls, parent, result, resultMeshes):
-		if parent.type == 'MESH':
-			resultMeshes.append(len(result))
-		result.append(parent)
-
-		for c in parent.children:
-			ArmatureFromObjects.addChildren(c, result, resultMeshes)
-
-	def execute(self, context):
-		import mathutils
-
-		if len(context.selected_objects) == 0 or bpy.context.object.mode != 'OBJECT':
-			return {'CANCELLED'}
-		active = context.active_object
-
-		objects = list()
-		meshes = list()
-
-		ArmatureFromObjects.addChildren(active, objects, meshes)
-
-		zfillSize = max(2, len(str(len(meshes))))
-
-		if len(objects) == 1:
-			return {'CANCELLED'}
-
-		armature = bpy.data.armatures.new("ARM_" + active.name)
-		armatureObj = bpy.data.objects.new("ARM_" + active.name, armature)
-		armatureObj.parent = active.parent
-		armatureObj.matrix_local = active.matrix_local
-		globalMatrix = active.matrix_world
-
-		context.scene.collection.objects.link(armatureObj)
-
-		bpy.ops.object.select_all(action='DESELECT')
-		context.view_layer.objects.active = armatureObj
-		bpy.ops.object.mode_set(mode='EDIT', toggle=False)
-
-		edit_bones = armatureObj.data.edit_bones
-		boneMap = dict()
-		bones = objects[1:]
-
-		scales = dict()
-
-		for b in bones:
-			boneName = b.name
-			bone = edit_bones.new(b.name)
-			bone.layers[0] = True
-			bone.head = (0,0,0)
-			bone.tail = (1,0,0)
-
-			pos, rot, _ = b.matrix_world.decompose()
-
-			matrix = mathutils.Matrix.Translation(pos) @ rot.to_matrix().to_4x4()
-
-			bone.matrix = globalMatrix.inverted() @ matrix
-
-			if b.parent in bones:
-				bone.parent = boneMap[b.parent]
-			boneMap[b] = bone
-
-			_, _, poseScale = b.matrix_local.decompose()
-			if poseScale.x != 1 or poseScale.y != 1 or poseScale.z != 1:
-				scales[boneName] = poseScale
-
-		bpy.ops.object.mode_set(mode='OBJECT')
-		meshCount = 0
-
-		print("making meshes")
-
-		bpy.ops.object.select_all(action='DESELECT')
-
-		meshOBJs = list()
-
-		for i in meshes:
-			boneObject = objects[i]
-
-			meshCopy = boneObject.data.copy()
-			meshCopy.name = "Mesh_" + str(meshCount).zfill(zfillSize)
-			meshObj = boneObject.copy()
-			meshObj.name = meshCopy.name
-			meshObj.data = meshCopy
-			context.scene.collection.objects.link(meshObj)
-
-			meshObj.parent = armatureObj
-			meshObj.matrix_local = globalMatrix.inverted() @ boneObject.matrix_world
-
-			bpy.ops.object.mode_set(mode='OBJECT')
-			meshObj.select_set(True, view_layer=context.view_layer)
-			meshObj.scale = (1,1,1)
-			bpy.ops.object.transform_apply(location = True, scale = True, rotation = True)
-
-			modif = meshObj.modifiers.new("deform", 'ARMATURE')
-			modif.object = armatureObj
-
-			group = meshObj.vertex_groups.new(name=boneObject.name)
-			group.add([v.index for v in meshCopy.vertices], 1, 'ADD')
-
-			meshCount += 1
-			meshObj.select_set(False, view_layer=context.view_layer)
-
-			meshOBJs.append(meshObj)
-
-		if self.mergeModel:
-			for o in meshOBJs:
-				o.select_set(True)
-			context.view_layer.objects.active = meshOBJs[0]
-			bpy.ops.object.join()
-
-		bpy.ops.object.select_all(action='DESELECT')
-
-		print("made meshes")
-
-		armatureObj.rotation_mode = self.rotMode
-		for b in armatureObj.pose.bones:
-			b.rotation_mode = self.rotMode
-			if b.name in scales:
-				b.scale = scales[b.name]
-
-
-		return {'FINISHED'}
-
-class AddTextureSlot(bpy.types.Operator):
-	bl_idname = "scene.saaddtexturesslot"
-	bl_label="Add texture"
-	bl_description="Adds texture to the texture list"
-
-
-	def execute(self, context):
-		settings = context.scene.saSettings
-
-		# getting next usable global id
-		ids = list()
-		for t in settings.textureList:
-			ids.append(t.globalID)
-		ids.sort(key=lambda x: x)
-		globalID = -1
-		for i, index in enumerate(ids):
-			if i != index:
-				globalID = i
-				break
-		if globalID == -1:
-			globalID = len(settings.textureList)
-
-		# creating texture
-		tex = settings.textureList.add()
-		#tex.image_user
-		settings.active_texture_index = len(settings.textureList) -1
-
-		tex.name = "Texture"
-		tex.globalID = globalID
-
-		return {'FINISHED'}
-
-class RemoveTextureSlot(bpy.types.Operator):
-	bl_idname = "scene.saremovetexturesslot"
-	bl_label="Remove texture"
-	bl_description="Removes the selected texture from the texture list"
-
-	def execute(self, context):
-		settings = context.scene.saSettings
-		settings.textureList.remove(settings.active_texture_index)
-
-		settings.active_texture_index -= 1
-		if len(settings.textureList) == 0:
-			settings.active_texture_index = -1
-		elif settings.active_texture_index < 0:
-			settings.active_texture_index = 0
-
-		return {'FINISHED'}
-
-class MoveTextureSlot(bpy.types.Operator):
-	bl_idname = "scene.samovetexturesslot"
-	bl_label="Move texture"
-	bl_description="Moves texture slot in list"
-
-	direction: EnumProperty(
-		name="Direction",
-		items=( ('UP',"up","up"),
-				('DOWN',"down","down"), )
-		)
-
-	def execute(self, context):
-		settings = context.scene.saSettings
-		newIndex = settings.active_texture_index + (-1 if self.direction == 'UP' else 1)
-		if not (newIndex == -1 or newIndex >= len(settings.textureList)):
-			if settings.correct_Material_Textures:
-				for m in bpy.data.materials:
-					props: SAMaterialSettings = m.saSettings
-					if props.b_TextureID == newIndex:
-						props.b_TextureID = settings.active_texture_index
-					elif props.b_TextureID == settings.active_texture_index:
-						props.b_TextureID = newIndex
-
-			settings.textureList.move(settings.active_texture_index, newIndex)
-
-			settings.active_texture_index = newIndex
-		return {'FINISHED'}
-
-class ClearTextureList(bpy.types.Operator):
-	bl_idname = "scene.sacleartexturelist"
-	bl_label="Clear list"
-	bl_description="Removes all entries from the list"
-
-	def execute(self, context):
-		settings = context.scene.saSettings
-		settings.active_texture_index = -1
-
-		for t in settings.textureList:
-			if t.image is not None:
-				t.image.use_fake_user = False
-
-
-		settings.textureList.clear()
-		return {'FINISHED'}
-
-class AutoNameTextures(bpy.types.Operator):
-	bl_idname = "scene.saautonametexlist"
-	bl_label="Autoname entries"
-	bl_description="Renames all entries to the assigned texture"
-
-	def execute(self, context):
-		texList = context.scene.saSettings.textureList
-
-		for t in texList:
-			if t.image is not None:
-				t.name = os.path.splitext(t.image.name)[0]
-			else:
-				t.name = "Texture"
-		return {'FINISHED'}
-
-def autotexUpdate(context, newValue):
-	# variables
-	mats = []
-	copyTex = []
-	index = -1
-	found = False
-
-	# collections
-	objects = context.visible_objects
-	texList = context.scene.saSettings.textureList
-
-	# place texList into array for ease of reference
-	for t in texList:
-		#print("Process Texlist")
-		if t.image is not None:
-			copyTex.append(t.name)
-
-	# loop through all objects
-	for o in objects:
-		#print("Checking Objects")
-		if o.type == 'MESH':
-			# if mesh type, check material slots
-			for matslot in o.material_slots:
-				if matslot.material:
-					matProps : SAMaterialSettings = matslot.material.saSettings
-					nodes = matslot.material.node_tree.nodes
-					for n in nodes:
-						# search valid materials for Texture Image input node
-						if n.type == 'TEX_IMAGE':
-							matname = os.path.splitext(str(n.image.name))[0]
-							if matname in texList:
-								#print("Found, " + matname)
-								index = copyTex.index(matname)
-								if index > -1:
-									# if updated index is > -1, update SA Material Texture ID
-									matProps.b_TextureID = index
-									#print("Texture updated!")
-									found = True
-			
-			# if Texture Image input cannot be found, check material name as backup.
-			if found == False:
-				for matslot in o.material_slots:
-					if matslot.material:
-						matname = matslot.material.name
-						if matname in texList:
-							#print("Found, " + matname)
-							index = copyTex.index(matname)
-							if index > -1:
-								# if updated index > -1, update SA Material Texture ID
-								matProps.b_TextureID = index
-								#print("Texture updated!")
-						else:
-							print("Error! No matching texture for " + matname + "in object " + o.name)
-
-class AutoAssignTextures(bpy.types.Operator):
-	bl_idname = "scene.saautoassignmmd"
-	bl_label = "Auto Assign Textures"
-	bl_description = "Attempts to match texture or material names to an imported texlist. Only checks visible objects, may not work."
-
-	def execute(self, context):
-		autotexUpdate(context, True)
-		return {'FINISHED'}
-
-class ToPrincipledBsdf(bpy.types.Operator):
-	bl_idname = "scene.convmaterials"
-	bl_label = "Convert to Principled BSDF"
-	bl_description = "Converts materials to a Principled BSDF for exporting purposes."
-
-	def execute(self, context):
-		for m in bpy.data.materials:
-			tree: bpy.types.NodeTree = m.node_tree
-			nodes = tree.nodes
-			out = nodes.get('Material Output')
-			saShader = nodes.get('Group', None)
-			texture = nodes.get('Image Texture', None)
-			bsdf = nodes.new('ShaderNodeBsdfPrincipled')
-			bsdf.inputs[5].default_value = 0
-
-			if saShader is not None:
-				nodes.remove(saShader)
-
-			if texture is not None:
-				tree.links.new(texture.outputs[0], bsdf.inputs[0]) # Assign texture
-				tree.links.new(texture.outputs[1], bsdf.inputs[21]) # Assign alpha
-
-			tree.links.new(bsdf.outputs[0], out.inputs[0])
-		return {'FINISHED'}
-
-class UpdateMaterials(bpy.types.Operator):
-	bl_idname = "scene.saupdatemats"
-	bl_label="Update Materials"
-	bl_description="Sets material nodetrees and variables of all selected objects to imitate how they would look in sadx/sa2"
-
-	@classmethod
-	def addDriver(cls, inputSocket, scene, path, entry = -1):
-		#curve = inputSocket.driver_add("default_value")
-		#driver = curve.driver
-		#driver.type = 'AVERAGE'
-		#variable = driver.variables.new()
-		#variable.targets[0].id_type = 'SCENE'
-		#variable.targets[0].id = scene
-		#variable.targets[0].data_path = "saSettings." + path + ("" if entry == -1 else "[" + str(entry) + "]")
-		#curve.update()
-		inputSocket.default_value = getattr(scene.saSettings, path) if entry == -1 else getattr(scene.saSettings, path)[entry]
-
-	def execute(self, context):
-		# remove old trees
-
-		ng = bpy.data.node_groups
-
-		n = ng.find("UV Tiling")
-		if n > -1:
-			ng.remove(ng[n])
-
-		n = ng.find("SAShader")
-		if n > -1:
-			ng.remove(ng[n])
-
-		n = ng.find("EnvMap")
-		if n > -1:
-			ng.remove(ng[n])
-
-		# now reload them them
-		directory = os.path.dirname(os.path.realpath(__file__)) + "\\Shaders.blend\\NodeTree\\"
-		bpy.ops.wm.append(filename="UV Tiling", directory=directory)
-		bpy.ops.wm.append(filename="SAShader", directory=directory)
-		bpy.ops.wm.append(filename="EnvMap", directory=directory)
-
-		tilingGroup = ng[ng.find("UV Tiling")]
-		saShaderGroup = ng[ng.find("SAShader")]
-		envMapGroup = ng[ng.find("EnvMap")]
-
-		# Drivers dont update automatically when set like this, and i cant find a way to update them through python, so we'll just set them temporarily
-		nd = saShaderGroup.nodes
-		lightDirNode: bpy.types.ShaderNodeCombineXYZ = nd[nd.find("LightDir")]
-		lightColNode: bpy.types.ShaderNodeCombineRGB = nd[nd.find("LightCol")]
-		ambientColNode: bpy.types.ShaderNodeCombineRGB = nd[nd.find("AmbientCol")]
-		displSpecularNode: bpy.types.ShaderNodeValue = nd[nd.find("DisplSpecular")]
-
-		UpdateMaterials.addDriver(lightDirNode.inputs[0], context.scene, "LightDir", 0)
-		UpdateMaterials.addDriver(lightDirNode.inputs[1], context.scene, "LightDir", 1)
-		UpdateMaterials.addDriver(lightDirNode.inputs[2], context.scene, "LightDir", 2)
-
-		UpdateMaterials.addDriver(lightColNode.inputs[0], context.scene, "LightColor", 0)
-		UpdateMaterials.addDriver(lightColNode.inputs[1], context.scene, "LightColor", 1)
-		UpdateMaterials.addDriver(lightColNode.inputs[2], context.scene, "LightColor", 2)
-
-		UpdateMaterials.addDriver(ambientColNode.inputs[0], context.scene, "LightAmbientColor", 0)
-		UpdateMaterials.addDriver(ambientColNode.inputs[1], context.scene, "LightAmbientColor", 1)
-		UpdateMaterials.addDriver(ambientColNode.inputs[2], context.scene, "LightAmbientColor", 2)
-
-		UpdateMaterials.addDriver(displSpecularNode.outputs[0], context.scene, "DisplaySpecular")
-
-		# The materials know whether the shader displays vertex colors based on the object color, its weird i know, but i didnt find any better way
-		import math
-		for o in context.scene.objects:
-			if o.type == 'MESH':
-				isNrm = o.data.saSettings.sa2ExportType == 'NRM'
-				r = o.color[0]
-				rc = bool(math.floor(r * 1000) % 2)
-
-				if isNrm and rc:
-					r = (math.floor(r * 1000) + (1 if r < 1 else (-1))) / 1000.0
-				elif not isNrm and not rc:
-					r = (math.floor(r * 1000) + ((-1) if r > 0 else 1)) / 1000.0
-				o.color[0] = r
-
-		# now its time to set all of the materials
-		for m in bpy.data.materials:
-			#creating an settings nodes
-			mProps = m.saSettings
-			m.use_nodes = True
-			tree: bpy.types.NodeTree = m.node_tree
-			nodes = tree.nodes
-			nodes.clear()
-
-			out = nodes.new("ShaderNodeOutputMaterial")
-			saShader = nodes.new("ShaderNodeGroup")
-			saShader.node_tree = saShaderGroup
-
-			saShader.inputs[2].default_value = mProps.b_Diffuse
-			saShader.inputs[3].default_value = mProps.b_Diffuse[3]
-			saShader.inputs[4].default_value = mProps.b_ignoreLighting
-
-			saShader.inputs[5].default_value = mProps.b_Specular
-			saShader.inputs[6].default_value = mProps.b_Specular[3]
-			saShader.inputs[7].default_value = mProps.b_Exponent
-			saShader.inputs[8].default_value = mProps.b_ignoreSpecular
-
-			saShader.inputs[9].default_value = mProps.b_Ambient
-			saShader.inputs[10].default_value = mProps.b_Ambient[3]
-			saShader.inputs[11].default_value = mProps.b_ignoreAmbient
-
-			saShader.inputs[12].default_value = mProps.b_flatShading
-
-			saShader.location = (-200, 0)
-
-			tree.links.new(out.inputs[0], saShader.outputs[0])
-
-			if mProps.b_useTexture:
-				try:
-					img = context.scene.saSettings.textureList[mProps.b_TextureID]
-				except IndexError:
-					img = None
-
-				if img is not None:
-					tex = nodes.new("ShaderNodeTexImage")
-					tex.image = img.image
-					tex.interpolation = 'Closest' if mProps.b_texFilter == 'POINT' else 'Smart'
-					tex.location = (-500, 0)
-					tree.links.new(tex.outputs[0], saShader.inputs[0])
-					tree.links.new(tex.outputs[1], saShader.inputs[1])
-
-
-					uvNode = nodes.new("ShaderNodeGroup")
-					if mProps.b_useEnv:
-						uvNode.node_tree = envMapGroup
-					else:
-						uvNode.node_tree = tilingGroup
-
-						uvNode.inputs[1].default_value = mProps.b_mirrorV
-						uvNode.inputs[2].default_value = mProps.b_mirrorU
-						uvNode.inputs[3].default_value = mProps.b_clampV
-						uvNode.inputs[4].default_value = mProps.b_clampU
-
-						uvSrc = nodes.new("ShaderNodeUVMap")
-						uvSrc.location = (-900, 0)
-						tree.links.new(uvSrc.outputs[0], uvNode.inputs[0])
-					tree.links.new(uvNode.outputs[0], tex.inputs[0])
-					uvNode.location = (-700, 0)
-				else:
-					saShader.inputs[0].default_value = (1,0,1,1)
-
-			if mProps.b_useAlpha:
-				m.blend_method = context.scene.saSettings.viewportAlphaType
-				m.alpha_threshold = context.scene.saSettings.viewportAlphaCutoff
-			else:
-				m.blend_method = 'OPAQUE'
-			m.shadow_method = 'NONE'
-			m.use_backface_culling = not mProps.b_doubleSided
-
-		# setting the color management
-		context.scene.display_settings.display_device = 'sRGB'
-		context.scene.view_settings.view_transform = 'Standard'
-		context.scene.view_settings.look = 'None'
-		context.scene.view_settings.exposure = 0
-		context.scene.view_settings.gamma = 1
-		context.scene.sequencer_colorspace_settings.name = 'sRGB'
-		context.scene.view_settings.use_curve_mapping = False
-		return {'FINISHED'}
-
-class MatToAssetLibrary(bpy.types.Operator):
-	bl_idname = "scene.samatslib"
-	bl_label="Materials to Asset Library"
-	bl_description="Generates new materials from an imported texlist and automatically marks them for the Asset Library."
-
-	# uncomment print commands in case of debugging.
-	def execute(self, context):
-		texList = context.scene.saSettings.textureList
-		arr_tls = []
-		genmats = []
-
-		# no apparent method for retrieving the index of a bpy_collection, so we use an array
-		for t in texList:
-			arr_tls.append(t.name)
-
-		for tex in texList:
-			idx = 0
-			texname = tex.name
-			
-			if tex.image is not None:
-				matname = texname
-			else:
-				matname = "material_" + str(arr_tls.index(texname))
-
-			if matname not in bpy.data.materials:
-				#print("Material Name Not Found, Creating new material.")
-				mat = bpy.data.materials.new(name=matname)
-			else:
-				#print("Material Name Found, Updating exisitng material.")
-				mat = bpy.data.materials[matname]
-
-			matProps : SAMaterialSettings = mat.saSettings
-
-			# set bpy.data.materials specifics here
-			mat.preview_render_type = 'FLAT'	## sets preview image to flat preview, better for viewing textures
-
-			# set saMaterialSettings specifics here
-			matProps.b_TextureID = arr_tls.index(texname)	## updates saSettings Texture ID for material
-
-			genmats.append(mat)
-
-		
-		UpdateMaterials.execute(self, context)
-
-		for m in genmats:
-			# add to asset library after updating material
-			m.asset_mark()					## marks as asset for asset library
-			m.asset_generate_preview()		## generates preview image in the asset library
-
-		return {'FINISHED'}
 
 #  quick edit
 def qeUpdate(context, newValue):
